Re-write Ruby Style Guide
On reviewing this page it seemed to provide very little general Ruby
information and lots of specific RSpec information. This felt a little
lopsided. I've now re-written this to have a heavier emphasis on
general, broad Ruby conventions.

The RSpec documentation has been removed as it seemed an odd fit for GDS
Way and felt like it exposed more of the GOV.UK programmes preferences
than it did GDS conventions. Some of the information in it was
out-of-date, for example the example given, with it's use of subject,
would flag as a violation with rubocop-govuk. I've instead included a
single paragraph to explain the preference for RSpec.

I've included links to some GOV.UK Ruby documentation that should be
helpful for further Ruby understanding. The Rails conventions
documentation covers a variety of the same areas as the RSpec
documentation did so this information should remain accessible.